Bitte mal anschauen. Beide datas sehen gleich aus, gleich klasser gleicher aufbau. warum funktioniert das ganze nur mit der dataMap und nicht mit JSON?
Data:
Data read from file: [
  { name: 'User 1', connections: [ 1, 2, 3 ] },
  { name: 'User 2', connections: [ 0, 2 ] },
  { name: 'User 3', connections: [ 0, 1, 3, 4, 5, 6 ] },
  { name: 'User 4', connections: [ 0, 2, 4, 5, 6, 7 ] },
  { name: 'User 5', connections: [ 2, 3, 5, 6, 7 ] },
  { name: 'User 6', connections: [ 2, 3, 4, 6, 7 ] },
  { name: 'User 7', connections: [ 2, 3, 4, 5, 7 ] },
  { name: 'User 8', connections: [ 4, 5, 6 ] },
  { name: 'User 9', connections: [ 4, 5, 6 ] }
]
