+++ script.js
@@ -1,15 +1,19 @@
-/*const fs = require('fs');
+const fs = require('fs');
 const path = require('path');
 
-// Function to read data from JSON file and convert to Map
+// Function to read data from JSON file and convert to array of objects
 function readDataFromFile(filename) {
     try {
         const jsonData = JSON.parse(fs.readFileSync(filename, 'utf-8'));
-        const data = new Map();
-        Object.keys(jsonData).forEach(key => {
-            data.set(key, jsonData[key]);
+        const dataArray = [];
+
+        // Iterate through each key-value pair in the JSON data
+        Object.keys(jsonData).forEach(name => {
+            const connections = jsonData[name].map(targetName => Object.keys(jsonData).indexOf(targetName));
+            dataArray.push({ name, connections });
         });
-        return data;
+
+        return dataArray;
     } catch (error) {
         console.error(`Error reading data from ${filename}: ${error.message}`);
         return null;
@@ -19,40 +23,9 @@ function readDataFromFile(filename) {
 // Get the absolute path of data.json using path.join
 const filename = path.join(__dirname, 'data.json');
 
-// Test reading data from JSON file
-const dataFromFile = readDataFromFile(filename);
-console.log("Data read from file:", dataFromFile);*/
-
-function fillData() {
-    const data = new Map();
-    data.set("User 1", ["User 2", "User 3", "User 4"]);
-    data.set("User 2", ["User 1", "User 3"]);
-    data.set("User 3", ["User 1", "User 2", "User 4", "User 5", "User 6", "User 7"]);
-    data.set("User 4", ["User 1", "User 3", "User 5", "User 6", "User 7", "User 8"]);
-    data.set("User 5", ["User 3", "User 4", "User 6", "User 7", "User 8"]);
-    data.set("User 6", ["User 3", "User 4", "User 5", "User 7", "User 8"]);
-    data.set("User 7", ["User 3", "User 4", "User 5", "User 6", "User 8"]);
-    data.set("User 8", ["User 5", "User 6", "User 7"]);
-    data.set("User 9", ["User 5", "User 6", "User 7"]);
-    return data;
-}
-
-// Function to convert the map to an array of objects
-function mapToArray(dataMap) {
-    const dataArray = [];
-    const nodeNames = Array.from(dataMap.keys());
-
-    nodeNames.forEach((name, index) => {
-        const connections = dataMap.get(name).map(targetName => nodeNames.indexOf(targetName));
-        dataArray.push({ name, connections });
-    });
-
-    return dataArray;
-}
-
-// Generate the data
-const dataMap = fillData();
-const data = mapToArray(dataMap);
+// Test reading data from JSON file and converting to array of objects
+const data = readDataFromFile(filename);
+console.log("Data read from file:", data);
 
 // Create SVG
 const svg = d3.select("#graph");
